# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders": {
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrdersDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/export/excel": {
      "get": {
        "operationId": "OrdersController_exportExcel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Export filtered orders as Excel"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/manual/products": {
      "get": {
        "operationId": "OrdersController_searchManualProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Search products for manual order creation"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/manual": {
      "post": {
        "operationId": "OrdersController_createManual",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManualOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create manual order from admin panel"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/manual/edit": {
      "post": {
        "operationId": "OrdersController_updateManual",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateManualOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update manual order before shipping"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}": {
      "get": {
        "operationId": "OrdersController_getById",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/confirm": {
      "post": {
        "operationId": "OrdersController_confirm",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/cancel": {
      "post": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/complete": {
      "post": {
        "operationId": "OrdersController_complete",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/fulfillment/start-preparing": {
      "post": {
        "operationId": "OrdersController_startPreparing",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/fulfillment/start-preparing-with-payment-override": {
      "post": {
        "operationId": "OrdersController_overrideStartPreparing",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
    "/orders/{orderId}/fulfillment/dispatch": {
      "post": {
        "operationId": "OrdersController_dispatch",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/fulfillment/mark-fulfilled": {
      "post": {
        "operationId": "OrdersController_markFulfilled",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/fulfillment/mark-failed": {
      "post": {
        "operationId": "OrdersController_markFailed",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/fulfillment/retry": {
      "post": {
        "operationId": "OrdersController_retry",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/orders/{orderId}/fulfillment/cancel": {
      "post": {
        "operationId": "OrdersController_cancelFulfillment",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
    "/payments/order/{orderId}": {
      "get": {
        "operationId": "PaymentsController_getByOrderId",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get payment by order ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "General Ecommerce API",
    "description": "General Ecommerce backend APIs",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FulfillmentSummaryDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "delivery",
              "pickup",
              "external_shipping",
              "manual_coordination"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "unfulfilled",
              "preparing",
              "ready",
              "out_for_delivery",
              "fulfilled",
              "failed",
              "cancelled"
            ]
          },
          "statusLabel": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "status",
          "statusLabel"
        ]
      },
      "MoneyTotalsDto": {
        "type": "object",
        "properties": {
          "subtotalAmount": {
            "type": "string"
          },
          "discountAmount": {
            "type": "string"
          },
          "shippingAmount": {
            "type": "string"
          },
          "taxAmount": {
            "type": "string"
          },
          "totalAmount": {
            "type": "string"
          },
          "paidAmount": {
            "type": "string"
          },
          "refundedAmount": {
            "type": "string"
          },
          "refundableAmount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        },
        "required": [
          "subtotalAmount",
          "discountAmount",
          "shippingAmount",
          "taxAmount",
          "totalAmount",
          "paidAmount",
          "refundedAmount",
          "refundableAmount",
          "currency"
        ]
      },
      "CustomerSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "phone"
        ]
      },
...
      "OrderSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "confirmed",
              "completed",
              "cancelled"
            ]
          },
          "statusLabel": {
            "type": "string"
          },
          "fulfillment": {
            "$ref": "#/components/schemas/FulfillmentSummaryDto"
          },
          "totals": {
            "$ref": "#/components/schemas/MoneyTotalsDto"
          },
          "version": {
            "type": "number"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummaryDto"
          },
          "paymentSummary": {
            "$ref": "#/components/schemas/PaymentSummaryDto"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "statusLabel",
          "fulfillment",
          "totals",
          "version",
          "note",
          "createdAt",
          "updatedAt",
          "customer",
          "paymentSummary"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "OrderListSummaryDto": {
        "type": "object",
        "properties": {
          "statusCounts": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          }
        },
        "required": [
          "statusCounts"
        ]
      },
      "PaginatedOrdersDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderSummaryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          },
          "summary": {
            "$ref": "#/components/schemas/OrderListSummaryDto"
          }
        },
        "required": [
          "data",
          "meta",
          "summary"
        ]
      },
      "CreateManualOrderDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateManualOrderDto": {
        "type": "object",
        "properties": {}
      },
...
      "OrderTimelineDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "nullable": true
          },
          "to": {
            "type": "string"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "from",
          "to",
          "note",
          "createdAt"
        ]
      },
...
      "OrderDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "confirmed",
              "completed",
              "cancelled"
            ]
          },
          "statusLabel": {
            "type": "string"
          },
          "fulfillment": {
            "$ref": "#/components/schemas/FulfillmentSummaryDto"
          },
          "totals": {
            "$ref": "#/components/schemas/MoneyTotalsDto"
          },
          "version": {
            "type": "number"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummaryDto"
          },
          "paymentSummary": {
            "$ref": "#/components/schemas/PaymentSummaryDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemSnapshotDto"
            }
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderTimelineDto"
            }
          },
          "fulfillmentHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommercialHistoryDto"
            }
          },
          "paymentHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommercialHistoryDto"
            }
          },
          "inventoryReservations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryReservationSummaryDto"
            }
          },
          "auditTimeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommercialAuditTimelineDto"
            }
          },
          "payment": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderPaymentDto"
              }
            ]
          },
          "allowedTransitions": {
            "$ref": "#/components/schemas/AllowedTransitionsDto"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "statusLabel",
          "fulfillment",
          "totals",
          "version",
          "note",
          "createdAt",
          "updatedAt",
          "customer",
          "paymentSummary",
          "items",
          "timeline",
          "fulfillmentHistory",
          "paymentHistory",
          "inventoryReservations",
          "auditTimeline",
          "payment",
          "allowedTransitions"
        ]
      },
      "OrderCommandDto": {
        "type": "object",
        "properties": {}
      },
...
      "PaymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "submitted",
              "under_review",
              "approved",
              "rejected",
              "expired",
              "cancelled",
              "partially_refunded",
              "refunded"
            ]
          },
          "statusLabel": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "paidAmount": {
            "type": "string"
          },
          "refundedAmount": {
            "type": "string"
          },
          "refundableAmount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "submissionVersion": {
            "type": "number"
          },
          "storePaymentMethodId": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodCatalogId": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodCode": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodName": {
            "type": "string",
            "nullable": true
          },
          "accountName": {
            "type": "string",
            "nullable": true
          },
          "accountNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "iban": {
            "type": "string",
            "nullable": true
          },
          "instructionsAr": {
            "type": "string",
            "nullable": true
          },
          "instructionsEn": {
            "type": "string",
            "nullable": true
          },
          "payerReference": {
            "type": "string",
            "nullable": true
          },
          "payerReceiptUrl": {
            "type": "string",
            "nullable": true
          },
          "payerReceiptMediaAssetId": {
            "type": "string",
            "nullable": true
          },
          "payerNote": {
            "type": "string",
            "nullable": true
          },
          "customerSubmittedAt": {
            "type": "string",
            "nullable": true
          },
          "receiptUrl": {
            "type": "string",
            "nullable": true
          },
          "receiptMediaAssetId": {
            "type": "string",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "nullable": true
          },
          "reviewedBy": {
            "type": "string",
            "nullable": true
          },
          "reviewNote": {
            "type": "string",
            "nullable": true
          },
          "customerUploadedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "allowedTransitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllowedTransitionDto"
            }
          }
        },
        "required": [
          "id",
          "orderId",
          "method",
          "status",
          "statusLabel",
          "amount",
          "paidAmount",
          "refundedAmount",
          "refundableAmount",
          "currency",
          "version",
          "submissionVersion",
          "storePaymentMethodId",
          "paymentMethodCatalogId",
          "paymentMethodCode",
          "paymentMethodName",
          "accountName",
          "accountNumber",
          "phoneNumber",
          "iban",
          "instructionsAr",
          "instructionsEn",
          "payerReference",
          "payerReceiptUrl",
          "payerReceiptMediaAssetId",
          "payerNote",
          "customerSubmittedAt",
          "receiptUrl",
          "receiptMediaAssetId",
          "reviewedAt",
          "reviewedBy",
          "reviewNote",
          "customerUploadedAt",
          "createdAt",
          "updatedAt",
          "allowedTransitions"
        ]
      },
...
    }
  }
}